Move useful things to helpers, less noisy logging.
